+++ src/utils/camel_to_title.ts
@@ -1,0 +1,5 @@
+export const convertCamelToTitle = (text: string) => {
+  const result = text.replace(/([A-Z])/g, ' $1');
+  const finalResult = result.charAt(0).toUpperCase() + result.slice(1);
+  return finalResult;
+};
